For indexes we don't need ints, we can take advantage of ushort's. This will minimize the footprint by 2 on every calculation of fingerprints.

+++ src/SoundFingerprinting.Tests/Unit/Utils/FingerprintDescriptorTest.cs
@@ -52,7 +52,7 @@
         public void EncodeFingerprintWorksAsExpected()
         {
             float[] framesSpectrumPowers = new float[] { 2, 4, 8, 9, 1, 3, 5 };
-            int[] indexes = new[] { 3, 2, 6, 1, 5, 0, 4 };
+            ushort[] indexes = new[] { (ushort)3, (ushort)2, (ushort)6, (ushort)1, (ushort)5, (ushort)0, (ushort)4 };
             bool[] expected = new[] { false, false, false, false, true, false, true, false, false, false, false, false, false, false };
 
             bool[] encodedFingerprint = fingerprintDescriptor.EncodeFingerprint(framesSpectrumPowers, indexes, 2);
